+++ index.js
@@ -1,8 +1,0 @@
-'use strict';
-
-var injectStylesheet = require('./lib/inject-stylesheet');
-
-module.exports = {
-  injectWithAllowlist: function (styles, list) { return injectStylesheet(styles, list, true); },
-  injectWithBlocklist: function (styles, list) { return injectStylesheet(styles, list, false); }
-};