+++ CHANGELOG.md
@@ -1,10 +1,16 @@
+# unreleased
+
+_Breaking Changes_
+
+- drop IE8 support
+
 # 2.0.0
 
 _Breaking Changes_
 
-* change `injectWithBlacklist` to `injectWithBlocklist`
-* change `injectWithWhitelist` to `injectWithAllowlist`
+- change `injectWithBlacklist` to `injectWithBlocklist`
+- change `injectWithWhitelist` to `injectWithAllowlist`
 
 # 1.0.0
 
-* Initial release
+- Initial release