+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# unreleased
+# 4.0.0
 
 _Breaking Changes_